intergrated testing

+++ routes/genres.js
@@ -11,14 +11,13 @@ router.get("/", async (req, res) => {
 });
 
 router.get("/:id", validateObjectId, async (req, res) => {
-  try {
-    const genre = await Genres.findById(req.params.id);
-    res.send(genre);
-  } catch {
+  const genre = await Genres.findById(req.params.id);
+  if (!genre) {
     return res
       .status(404)
       .send(`Cannot find genres with given id : ${req.params.id}`);
   }
+  res.send(genre);
 });
 
 router.post("/", auth, async (req, res) => {
@@ -33,10 +32,16 @@ router.post("/", auth, async (req, res) => {
 
 router.put("/:id", auth, async (req, res) => {
   const { error } = validate(req.body);
-  if (error) return res.status(400).send(error.details[0].message);
+  if (error) return res.status(404).send(error.details[0].message);
 
-  const genre = await Genres.updateOne(
-    { _id: req.params.id },
+  const genreSearch = await Genres.findById(req.params.id);
+  if (!genreSearch) {
+    return res
+      .status(404)
+      .send(`Cannot find genres with given id : ${req.params.id}`);
+  }
+  const genre = await Genres.findByIdAndUpdate(
+    req.params.id,
     {
       name: req.body.name,
     },
@@ -46,12 +51,17 @@ router.put("/:id", auth, async (req, res) => {
 });
 
 router.delete("/:id", [auth, admin], async (req, res) => {
-  try {
-    const genre = await Genres.deleteOne({ _id: req.params.id });
-    res.send(genre);
-  } catch {
-    res.status(400).send(`Cannot find genres with id : ${req.params.id}`);
+  const genreSearch = await Genres.findById(req.params.id);
+  if (!genreSearch) {
+    return res
+      .status(404)
+      .send(`Cannot find genres with given id : ${req.params.id}`);
   }
+  const genre = await Genres.deleteOne({ _id: req.params.id });
+  if (!genre) {
+    res.status(404).send(`Cannot find genres with id : ${req.params.id}`);
+  }
+  res.send(genre);
 });
 
 module.exports = router;
